requirements.txt
Flask==2.2.2
werkzeug==2.2.2
gunicorn==20.1.0
requests==2.30.0
googlemaps==4.10.0
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
responses==0.24.1
google-generativeai
pytest-mock>=3.6.0